backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic>=2.9.0

# LangChain
langchain<0.3.0
langchain-community<0.3.0
langchain-text-splitters<0.3.0
langchain-chroma<0.2.0

# Vector DB
chromadb>=0.5.0

# PDF Parsing
pypdf>=4.3.0
requests>=2.32.0,<3.0.0

# LLM + Embeddings
groq>=0.5.0
sentence-transformers>=3.0.0
